py3.10-prep: Finally support Python 3.10
After the major problems are solved in
    py3.10-prep: re-implement zip import
    py3.10-prep: reset the type cache after feature switching
    py3.10-prep: Fix parser.py for changed typing module

, there are only some minor changes left to do. One thing
is still not clear:

* Unsolved: In time_test.py, the last two tests work but
crash at shutdown.
It is not yet clear if this is a PySide or Python error.

This will crash Python 3.10 if not solved until the release.

[ChangeLog][PySide6] Support Python 3.10. This is by 99% solved.
Looking for a crash with unknown origin.

Fixes: PYSIDE-1436
Change-Id: I94cffa7ed16a2651e09924fe5babc188b1b4c2b8
Reviewed-by: Christian Tismer <tismer@stackless.com>
(cherry picked from commit 81e9cff884d6b03cdf64a5dd6ec6564d99177e0c)
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>