MAYA-110705: Disable PySide2 __feature__ support, add comments.
This re-adds the __feature__ support, but in a separate function call,
and this function call is commented out, thus disabling the features.

A comment is left to users that see this, indicating they can uncomment
the call to add_pyside_feature_support() to re-enable access to
snake_case and true_import features of PySide2. Those features were
added in Qt 5.15.2, but really only advertised for use in Qt 6, so their
removal should not really be missed. They will return in Qt 6.