Jenkinsfile, All Builds: Update packaging for PySide6, update to 3.9.7
Update the Linux and macOS packaging scripts for PySide6.
Note that these are untested and just updated to PySide6 based on just
converting 2->6. That should cover most all of the changes, but the
layout might be a little different, so I do anticipate there will be
further changes to these scripts as they are tested.

Update Python version to 3.9.7 - the latest python artifacts that are
available in the team-maya-generic artifactory group.

Change-Id: I9d27fbda747986c5a0f5d96242d1e436708a966f