Windows Build: Add qtdir/bin to path, --qtpaths instead of --qtdir
Add Qt bin directory to the PATH to allow shiboken6 to properly build
Replace --qmake argument to setup.py with --qtpaths argument - as this
qmake is no longer used and the qmake option is deprecated.

Note that the change to add <qtdir>/bin to the PATH should not be
necessary in the future - as the need for this is a bug. Specifying
--qmake or --qtpaths on the setup.py command line should be enough to
allow setup.py to add <qtdir>/bin to the path for the build.
This bug is tracked as: https://bugreports.qt.io/browse/PYSIDE-1844

Change-Id: Ic8622c812f5ed355c661674927b3bda83f97d1e0