Jenkinsfile: Add support for branches that have been force-pushed
Make Jenkinsfile resilient to users who force-push a branch that Jenkins
has already built.

It does so by making sure it doesn't currently have the branch in
question checked out, deleting the branch, fetching, and then freshly
checking out the branch in question. Then the local branch is not at an
inconsistent state, and has the latest commits.

Change-Id: Idd15bc2b465b3b96d87a2a298da77a0b73a0aec9