Jenkinsfile: Update the artifact properties reflecting current versions
The artifact properties set when uploading an artifact include versions
for various dependencies - like the OS, compiler, and libclang version.
These were out of date/out of sync with the current OS, compiler and
libclang versions that are used, now updated.

Update macOS build config name to Monterey

The macOS build config name was not matching the OS version that it is
now building on. This updates the correct name and version reported.

Change-Id: Ic9f7daf801059c45c7d7cb1a1eb2f5dacaa748ba