Jenkinsfile: Correct auto-detection of Qt version to use
Prior to this change, QTVersion parameter had a default of 'latest',
which may be out of sync with the latest PySide6 version. Instead the
default should be the Qt version that matches the major.minor.patch
PySide6 version.

This changes QTVersion to instead show a list of choices, like is done
for the Python version. The default is to use 'match' which will try to
find the matching Qt version for this PySide6 version.

Change-Id: I6093d278a3b3bbe576b1ae018944af733a1dc8fe