Linux Packaging: Correct path to libpyside6.abi3.so.6.4.0
With these alpha/beta releases, the pyside version reports with an 'a1'
suffix (alpha 1) - e.g. 6.4.0a1. The alpha/beta suffix is not part of
the library name, so we need to strip that off when using the
PYSIDEVERSION var to refer to this library.

This also applies to pyside6qml and shiboken6 libraries.

Change-Id: I66b4d656001e6dd44440e35177cbf868a2dc50c2