Adapt Qt artifact fetch code for change in #artifacts generated
The Qt builds now generate 6 artifacts -- 2 for each platform - one for
the debuginfo files, the other for the rest of the installation.

The getQtArtifacts code needed to be adjusted so it was ok finding a
qt artifactory directory that contains 6 artifacts. Without this change,
it would skip the newer builds.

Change-Id: I1a032a8edfc157a39be13da46069e19a28f2216f