limited api: replace PySequence_Fast_GET_ITEM by PySequence_GetItem
PySequence_Fast_GET_ITEM is defined as:
    (PyList_Check(o) ? PyList_GET_ITEM(o, i) : PyTuple_GET_ITEM(o, i))

and when using the Limited API we re-define the _GET_ITEM macro
to be the _GetItem function, and considering this is our standard
use case, the macro could be replaced directly by the function.

However, the function returns a new reference, so we need to manually
drecrease a reference after the usage, to avoid reference counting
issues.

Change-Id: If361e80b9e40b033e009ad46b2b9430e5b4c8eaa
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
(cherry picked from commit fb13a26a76eba415e743d119d5d2782c607fee2f)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>