build: make FluentWinUI3 available for Linux and macOS
The style needs to be available for all supported
platforms, not only Windows.

We were still deploying the plugin, and when trying to find
the library for the style, it was affecting the deployment of
applications.

Change-Id: If8c0b08627beddf59cfb682cc2594ba06aca31f2
Fixes: PYSIDE-2910
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
(cherry picked from commit d42c7ccd72f6b818637c0ad7cfdbfd594ac7f95c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>