Fix smart pointer testing on macOS
The clang parser on macOS was not seeing namespace std for
std::optional. Add the system include <optional> to fix this.

This however unveils a problem with "T &std::optional::value()", which
returns a reference to the contained value instead of the value as
done by the synthesized function. The reference is handled as a
pointer conversion for object types and the std::optional itself
instead of the value is returned since it is found in the instance map
at the address (unhandled in findColocatedChild()).
Skip the test for the time being.

Task-number: PYSIDE-2854
Change-Id: Ifcb65643ff0a39c060fb329933a1e6949ed8777c
Reviewed-by: Christian Tismer <tismer@stackless.com>
(cherry picked from commit 4b68425a1102828f13c0f647fd0c6470577165ed)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>