shiboken6: Prepare the graph used for topological for sorting classes without operator==()
The graph used an array of Node to represent edges, which is
suitable only for small classes that have operator==().

Change the internal representation to use indexes for edges instead of
the node classes and add an API for adding edges by index. Rewrite the
visit() algorithm to create an index list instead of a node list. Move
the Node from NodeEntry to a separate node list, which allows for
copying it directly from the node list passed in.

Task-number: PYSIDE-2946
Task-number: PYSIDE-454
Change-Id: I7becd5caef226f50b77f5def2a62a2620d39ef8e
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>
(cherry picked from commit 3da05cc765ab16b3622e2262022a40aa92fc0334)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>