example_gallery: Fix Design Studio examples
Due to the Design Studio project structure, the doc .pyproject file is
located inside the Python folder. This change changes the project root
folder detection: instead of the parent of the .pyproject file, in the
case of DS projects the parent of the parent is the correct folder.

Change-Id: I54ba60dc9decf8bbcd53062fe02029568a8782a5
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
(cherry picked from commit e391b6a0192a21a96589721eba13a8eba3ad0c70)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>